move template method in abstract class
